+++ tests/empty.R
@@ -97,3 +97,9 @@ x = st_sf(a = 2:1, geometry = st_sfc(st_linestring(), st_linestring(matrix(1:4,2
 write_sf(x, "empty.shp")
 y = read_sf("empty.shp")
 all.equal(x, y)
+
+# https://github.com/edzer/sfr/issues/398 :
+pt = st_sfc(st_point(c(0,92)), crs = 4267)
+robin_crs <- "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
+india_crs <- "+init=epsg:24383"  # India-centered Lambert conformal conic projection
+st_transform(st_transform(pt, robin_crs), india_crs)